Make sure that sourcesets are not empty. Fixes eveoh/gradle-aspectj#4

+++ src/main/groovy/aspectj/AspectJPlugin.groovy
@@ -46,30 +46,35 @@ class AspectJPlugin implements Plugin<Project> {
             project.configurations.create('ajInpath')
         }
 
-        project.tasks.create(name: 'compileAspect', overwrite: true, description: 'Compiles AspectJ Source', type: Ajc) {
-            dependsOn project.configurations*.getTaskDependencyFromProjectDependency(true, "compileJava")
+        if (!project.sourceSets.main.allSource.isEmpty()) {
+            project.tasks.create(name: 'compileAspect', overwrite: true, description: 'Compiles AspectJ Source', type: Ajc) {
+                dependsOn project.configurations*.getTaskDependencyFromProjectDependency(true, "compileJava")
 
-            dependsOn project.processResources
-            sourceSet = project.sourceSets.main
-            inputs.files(sourceSet.allSource)
-            outputs.dir(sourceSet.output.classesDir)
-            aspectPath = project.configurations.aspectpath
-            ajInpath = project.configurations.ajInpath
+                dependsOn project.processResources
+                sourceSet = project.sourceSets.main
+                inputs.files(sourceSet.allSource)
+                outputs.dir(sourceSet.output.classesDir)
+                aspectPath = project.configurations.aspectpath
+                ajInpath = project.configurations.ajInpath
+            }
+
+            project.tasks.compileJava.deleteAllActions()
+            project.tasks.compileJava.dependsOn project.tasks.compileAspect
         }
-        project.tasks.compileJava.deleteAllActions()
-        project.tasks.compileJava.dependsOn project.tasks.compileAspect
 
+        if (!project.sourceSets.test.allSource.isEmpty()) {
+            project.tasks.create(name: 'compileTestAspect', overwrite: true, description: 'Compiles AspectJ Test Source', type: Ajc) {
+                dependsOn project.processTestResources, project.compileJava
+                sourceSet = project.sourceSets.test
+                inputs.files(sourceSet.allSource)
+                outputs.dir(sourceSet.output.classesDir)
+                aspectPath = project.configurations.aspectpath
+                ajInpath = project.configurations.ajInpath
+            }
 
-        project.tasks.create(name: 'compileTestAspect', overwrite: true, description: 'Compiles AspectJ Test Source', type: Ajc) {
-            dependsOn project.processTestResources, project.compileJava
-            sourceSet = project.sourceSets.test
-            inputs.files(sourceSet.allSource)
-            outputs.dir(sourceSet.output.classesDir)
-            aspectPath = project.configurations.aspectpath
-            ajInpath = project.configurations.ajInpath
+            project.tasks.compileTestJava.deleteAllActions()
+            project.tasks.compileTestJava.dependsOn project.tasks.compileTestAspect
         }
-        project.tasks.compileTestJava.deleteAllActions()
-        project.tasks.compileTestJava.dependsOn project.tasks.compileTestAspect
     }
 }
